server: Restore held entity to default state
During the GUADEC 25 workshop, we noticed that, when the holder entity
disconnects, the held entity remains in the HELD state.

This happens because the disconnect process does not specify the state
needed to update the held entity state.

Therefore, instead of simply specifying an arbitrary state, keep track
of the entity's default state and restore it when it's dropped.